fix: swith to float16

+++ run.py
@@ -78,7 +78,9 @@ def estimate_diabetes_probability(drugs: list, batch_size: int = 1) -> list:
 
 if __name__ == "__main__":
 
-    llm = LLM(model="meta-llama/Meta-Llama-3-8B-Instruct", tensor_parallel_size=2)
+    llm = LLM(model="meta-llama/Meta-Llama-3-8B-Instruct",
+              tensor_parallel_size=2,
+              dtype='float16')
     df = pd.read_parquet('drug_15355.parquet', engine='pyarrow')
     drugs = df['values'].tolist()[:100]
